+++ bridge-ui/src/controllers/PeginController.ts
@@ -37,14 +37,15 @@ export function setupSession(session: SessionInformation, setSession: React.Disp
   if (!session.isSet) {
     const sessionId = getCookie("sessionID");
     const escrowAddress = getCookie("escrowAddress");
+    const escrowScript = getCookie("escrowScript");
     const currentState = getCookie("currentState");
     const redeemAddress = getCookie("redeemAddress");
     const toplBridgePKey = getCookie("toplBridgePKey");
     const redeemTemplate = getCookie("redeemTemplate");
 
-    if (sessionId !== undefined && escrowAddress !== undefined && currentState !== undefined && redeemAddress !== undefined && toplBridgePKey !== undefined && redeemTemplate !== undefined) {
+    if (sessionId !== undefined && escrowAddress !== undefined && escrowScript !== undefined && currentState !== undefined && redeemAddress !== undefined && toplBridgePKey !== undefined && redeemTemplate !== undefined) {
       console.log("Session exists in cookie")
-      setSession({ isSet: true, sessionID: sessionId, escrowAddress: escrowAddress, currentState: stringToPeginUIState(currentState), redeemAddress: redeemAddress, toplBridgePKey: toplBridgePKey, redeemTemplate: redeemTemplate });
+      setSession({ isSet: true, sessionID: sessionId, escrowAddress: escrowAddress, escrowScript: escrowScript, currentState: stringToPeginUIState(currentState), redeemAddress: redeemAddress, toplBridgePKey: toplBridgePKey, redeemTemplate: redeemTemplate });
     }
   }
 }
@@ -52,28 +53,29 @@ export function setupSession(session: SessionInformation, setSession: React.Disp
 export function sessionStarted(setSession: React.Dispatch<React.SetStateAction<SessionInformation>>, response: StartSessionResponse) {
   setCookie("sessionID", response.sessionID);
   setCookie("escrowAddress", response.escrowAddress);
+  setCookie("escrowScript", response.script);
   setCookie("currentState", "PeginSessionStateWaitingForBTC");
-  setSession({ isSet: true, sessionID: response.sessionID, escrowAddress: response.escrowAddress, currentState: PeginUIState.SessionStarted, redeemAddress: "", toplBridgePKey: "", redeemTemplate: "" });
+  setSession({ isSet: true, sessionID: response.sessionID, escrowAddress: response.escrowAddress, escrowScript: response.script, currentState: PeginUIState.SessionStarted, redeemAddress: "", toplBridgePKey: "", redeemTemplate: "" });
 }
 
 export function btcArrived(setSession: React.Dispatch<React.SetStateAction<SessionInformation>>, session: SessionInformation) {
   setCookie("currentState", "PeginSessionStateMintingTBTC");
-  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, currentState: PeginUIState.MintingTBTC, redeemAddress: "", toplBridgePKey: "", redeemTemplate: "" });
+  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, escrowScript: session.escrowScript, currentState: PeginUIState.MintingTBTC, redeemAddress: "", toplBridgePKey: "", redeemTemplate: "" });
 }
 
 export function mintingBTC(setSession: React.Dispatch<React.SetStateAction<SessionInformation>>, session: SessionInformation) {
   setCookie("currentState", "PeginSessionStateMintingTBTC");
-  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, currentState: PeginUIState.MintingTBTC, redeemAddress: "", toplBridgePKey: "", redeemTemplate: "" });
+  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, escrowScript: session.escrowScript, currentState: PeginUIState.MintingTBTC, redeemAddress: "", toplBridgePKey: "", redeemTemplate: "" });
 }
 
 export function timeOutMintingBTC(setSession: React.Dispatch<React.SetStateAction<SessionInformation>>, session: SessionInformation) {
   setCookie("currentState", "PeginSessionTimeOutMintingTBTC");
-  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, currentState: PeginUIState.TimeOutMintingTBTC, redeemAddress: "", toplBridgePKey: "", redeemTemplate: "" });
+  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, escrowScript: session.escrowScript, currentState: PeginUIState.TimeOutMintingTBTC, redeemAddress: "", toplBridgePKey: "", redeemTemplate: "" });
 }
 
 export function timeOutBTCNotSent(setSession: React.Dispatch<React.SetStateAction<SessionInformation>>, session: SessionInformation) {
   setCookie("currentState", "PeginSessionTimeOutBTCNotSent");
-  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, currentState: PeginUIState.TimeOutBTCNotSent, redeemAddress: "", toplBridgePKey: "", redeemTemplate: "" });
+  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, escrowScript: session.escrowScript, currentState: PeginUIState.TimeOutBTCNotSent, redeemAddress: "", toplBridgePKey: "", redeemTemplate: "" });
 }
 
 
@@ -82,10 +84,10 @@ export function mintedBTC(setSession: React.Dispatch<React.SetStateAction<Sessio
   setCookie("redeemAddress", address);
   setCookie("toplBridgePKey", toplBridgePKey);
   setCookie("redeemTemplate", redeemTemplate);
-  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, currentState: PeginUIState.WaitingForRedemption, redeemAddress: address, toplBridgePKey: toplBridgePKey, redeemTemplate: redeemTemplate });
+  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, escrowScript: session.escrowScript, currentState: PeginUIState.WaitingForRedemption, redeemAddress: address, toplBridgePKey: toplBridgePKey, redeemTemplate: redeemTemplate });
 }
 
 export function claimedTBTC(setSession: React.Dispatch<React.SetStateAction<SessionInformation>>, session: SessionInformation) {
   setCookie("currentState", "PeginSessionWaitingForClaim");
-  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, currentState: PeginUIState.WaitingForClaim, redeemAddress: session.redeemAddress, toplBridgePKey: session.toplBridgePKey, redeemTemplate: session.redeemTemplate });
+  setSession({ isSet: true, sessionID: session.sessionID, escrowAddress: session.escrowAddress, escrowScript: session.escrowScript, currentState: PeginUIState.WaitingForClaim, redeemAddress: session.redeemAddress, toplBridgePKey: session.toplBridgePKey, redeemTemplate: session.redeemTemplate });
 }